refactor: interface 분리 작업
