+++ src/components/modal/modal_button/handlers.ts
@@ -1,7 +1,8 @@
 import RestaurantListStorageService from "../../../services/restaurantListStorageService";
 import validateRestaurantState from "../../../services/validateRestaurantState";
 import restaurantStateStore from "../../../store/RestaurantStateStore";
-import { IinvalidResult, Irestaurant, MappedType } from "../../../types";
+import { Irestaurant } from "../../../types/restaurant";
+import { IinvalidResult } from "../../../types/validate";
 import RestaurantList from "../../restaurant_list/RestaurantList";
 
 const initializeFormState = () => {
@@ -30,10 +31,7 @@ const removePrevErrorMessage = () => {
   });
 };
 
-const renderErrorMessage = (
-  index: number,
-  result: MappedType<IinvalidResult>,
-) => {
+const renderErrorMessage = (index: number, result: Partial<IinvalidResult>) => {
   const targetTag = document.getElementsByClassName("form-item")[index];
   const p = document.createElement("p");
   p.setAttribute("class", `invalid_message ${result.targetClassName}`);
@@ -41,7 +39,7 @@ const renderErrorMessage = (
   targetTag.appendChild(p);
 };
 
-const checkValidateHandler = (restaurantInfo: MappedType<Irestaurant>) => {
+const checkValidateHandler = (restaurantInfo: Partial<Irestaurant>) => {
   const validateResult = validateRestaurantState(restaurantInfo);
   removePrevErrorMessage();
