+++ src/domain/SetRestaurant.ts
@@ -1,4 +1,4 @@
-import { Irestaurant } from "../types";
+import { Irestaurant } from "../types/restaurant";
 
 export default function setRestaurant(restaurant: Irestaurant) {
   return {
