+++ src/types/category.ts
@@ -1,0 +1,8 @@
+export type Icategory = "한식" | "중식" | "일식" | "아시안" | "양식" | "기타";
+
+export type Iall = "전체";
+
+export interface IcategoryInfo {
+  category: string;
+  categoryImg: string;
+}
