+++ src/types/distance.ts
@@ -1,0 +1,1 @@
+export type Idistance = 5 | 10 | 15 | 20 | 30;
