+++ src/types/sort.ts
@@ -1,0 +1,1 @@
+export type IsortType = "name" | "distance";
