+++ src/types/validate.ts
@@ -1,0 +1,5 @@
+export type IinvalidResult = {
+  targetClassName: string;
+  isValid: boolean;
+  errorMessage: string;
+};
